# add ocl accleration for prior box layer and depthwise convolution