@vpisarev 2017-12-18
@alalek, I think, it would be nice to squeeze it into 3.4