@dkurt 2017-12-18
On my machine performance of MobileNet-SSD from Caffe (20 classes) is changed from 172.13ms to 26.66ms (22.71ms on CPU). MobileNet-SSD from TensorFlow for 90 classes is boosted from 203.47ms to 45.11ms (25.15ms on CPU). See https://github.com/opencv/opencv/wiki/DNN-Efficiency.

@pengli, May I aks you to add a test for the last model like a https://github.com/opencv/opencv/blob/master/modules/dnn/test/test_tf_importer.cpp#L189?