@pengli 2017-12-19
@dkurt,  I pushed one more commit  in this PR : "add ocl accuracy test for tf mobilenet ssd" ,  please check and see if this is what you want.