@pengli 2017-12-19
@alalek, convolution verification failure has been fixed in the new update.