@vpisarev 2017-12-19
@alalek, shall we merge it now? looks like the PR brings some significant performance improvements!